Moved Style to tkinter_style.py
